Adds support for np.array as input to ModelWrapper
It turns out to be messy to deal with tensor inputs such as ECG records
while still using a pandas dataframe as input. The solution would have
to use multi-index, but this gives further problems when doing the
cross-validation split. As a workaround, I now simply allow the input to
be a normal numpy array, and check the type wherever necessary. The
prediction output in the ModelWrapper is still turned into a dataframe.