Adds wrapper for tf.data.Dataset to extractor
The extractors should now return a Data object instead of the usual X,
y DataFrame. The Data object wraps an indexable datastructure and allows
it to be sliced without reading from the underlying data. The Data
object can be turned into either a numpy array or a tf.data.Dataset. The
extractor is expected to return a Container, which subclasses Data and
contains a nested set of Data objects in a dict-like structure. This
allows for a more flexible input to tensorflow models.

The main reason for making the new Data and Container objects is to
allow arbitrary slicing of the data without reading it into memory. This
makes it possible to perform custom cross-validation as usual without
having to actually read the whole dataset into memory first.