Reverts back to the old system with two conda envs
Apparently I was mistaken with not needing two versions of tensorflow
when using CPU or GPU, so I reverted back to the old "system", except
now GPU is default and there's a separate mim-cpu environment. Also had
to move jupyter_contrib_nbextensions to the pip-section of the yml file
for everything to work. But now it does work, I think, maybe.