Fixes slicing behavior of Data and Container
Adds some tests to illustrate desired behavior. Essentially, Data is
like a wrapper for some underlying array-like structure, but with the
twist that it maintains its own (hidden) mapping between the input index
and the index of the internal data. This mapping can be manipulated by
"slicing" the Data object. A slice of a slice should now work as
expected.

The Container class inherits Data, but with the added functionality
of being able to contain nested, dict-like objects.

This commit also updates the ECGData convenience class and moves it to
the extractor module.