Refactors cross-validation
A wrapper now allows regular sklearn cross-validators to be used as
expected. Adds a ChronologicalSplit "cross-validator" that simply slices
the data in two halves. This can also be used for holding out a dataset,
which is now added by default (although the hold-out size defaults to
0). A custom hold-out function could also be added.

This also refactors the initial MultipleECG experiments, which of course
have to be re-run after this.