Refactors experiments/run
Moves a lot of responsibility from the run module to the Experiment
class in preparation for hyper-experiments. Now, given an instance of
Experiment, you can do experiment.run() to run it.

Also changes the path to where experiment results are saved, including
where tensorboard logs are saved. Note that this will break old
experiment results.

Finally, some small fixes here and there, dealing with certain logging
messages and so on.