WIP, adds nice way to compare experiment params
Function to flatten a nested dictionary (such as the params of an
experiment), in a way that the keys are merged nicely. Another function
that can "stringify" any callables in a nested dict. Together, these two
functions allow me to save all the relevant arguments that define an
experiment in a future-proof dict that can be loaded even if the
underlying functions are moved or renamed. The experiment arguments are
now saved in the results file, and by flattening the dict the presenter
class can easily make a dataframe with one row per parameter, to allow
for easy comparison of arguments between different experiments.

The metadata report has been expanded to also store installed packages
in the current conda environment. Because this takes about a second to
retrieve, there's an option to toggle it off, which is used in the
tests to speed things up. The time cost is negligible when compared to
running actual experiments, however.

The presenter class has been changed so that it now loads experiments by
"globbing" the input string, rather than using the
Experiment.results_path to load the results. This makes it easier to
load results from experiments that are dynamically created, such as from
hyper-experiments. Also, it allows for loading results across different
derivations of the Experiment class, if we want to.