Optimization for datasets that fit in memory
Also moves hyper experiment search strategies to a separate file, and
adds a notebook that explores a hyper-experiment analyzing loss function
vs learning rate, batch size and dropout.

Streaming data from disk was a big bottleneck in my small models, so
I added a flag in Data for telling if the data fits in RAM. If it does
(and the flag is on) then the dataset is backed by numpy arrays rather
than my custom generator. This is much faster when reading the ECGs from
memory.