Removes DataProvider in favor of PredefinedSplits
The Data object now comes with an optional attribute that defines
predefined splits to use for making cross-validation splits (or just
regular train/val split as a special case). The Extractor class is
responsible for defining the splits to use. If the Data has predefined
splits, we can then set a flag that will use those splits instead of
regular cross-validation.