Adds experiment R1_TUNED_D100
The idea is that the "latent space" of the first dense layer after the
conv layers might have to be larger than 10 to be able to capture
fine-grained details of the ECG. So if we increase this layer size, the
models that combine multiple ECGs might be better able to spot
meaningful differences.

This commit also adds a WIP for defining the MACE labels.