Refactors serial ecg experiments and extractor
This is a breaking, unfinished change. The experiments will not run
properly here.

Additional work on esc-trop extractor and massage, for calculating MACE
labels etc.

Adds a column parameter to Data class, that can be used to save metadata
about which features are being prepared by an extractor. Useful for
saving experiment results later. If extractors also add an 'index'
variable in addition to 'x' and 'y', it can be used to debug the data,
see ESCTropExtractor for an example.