Major refactorings of ESC-Trop experiments
Renames all the experiments and removes some of the less interesting
ones. I tried to consolidate the different variations of cnn-models into
a single model now called ecg_cnn - still need to clean up and remove
unused models.

Adds a pre-processing step after the data-split in Experiment, allowing
an arbitrary function to modify the data before training starts on that
split. In particular, when using pre-trained models for ECG inputs, this
should give some speedup as the data is only processed once.

The Data class now has an optional column-parameter that helps keeping
track of feature names for nested data structures. Can be used together
with an index to make it easier to backtrack and look at individual
examples - see EscTrop extractor for example.

Changes (again) the calculation of MACE: some deaths were not caught due
to ambiguous timestamps - this has now been fixed so that when the day
of death is unknown, if the month could have been within 30 days, it
still counts towards MACE. When month of death is unknown, death only
counts towards MACE if the index visit was in november or december of
that year (there are no such cases in the ESC-Trop data, though). This
change also fixes a handful of deaths that were missed because they
occurred on the same day as the index (the delta-t was negative).

The 1010 patients with missing visit causes are now dropped from the
index.

Some of the experiments may have been changed subtly, but they will all
have to be re-run after this change anyway.