Adds Data/Container function to flatten data
The as_flat_numpy parameter will return a flattened version of the
underlying data, as a numpy array. For example, if the data is a tensor
with dimensions (n, 1200, 12), then the flattened data will be reshaped
to (n, 14400). For nested structures (Containers), each dict item will
first be flattened, then all the flattened items are concatenated,
maintaining the first dimension of the array.

The reason I need this flattening stuff is for training classifiers like
random forest which expect a simple 2-dimensional X-matrix as input. But
with the extractor structure I have right now, some of the input data
might be nested within a dictionary, like this: x = {'flat_features':
..., 'ecg_0': ...}. Even when I only have one item in the dict, RF will
be confused, so I want an easy way to just flatten all of it before
plugging it into the classifier, and the above should do just that.

Incidentally, this fixes (I think) the recently added, non-functional
RF-experiments.