Adds a bunch of logistic regression experiments
This adds a logistic_regression model implemented in keras as
a neural-network with a single layer. For this to work properly, set
batch-size to -1, use SGD optimizer with a learning-rate of 1, and make
sure to normalize the input. I use batch-normalization here, hopefully
that is equivalent since the batch-size is 1.