Adds shuffle-setting to Data.as_dataset
When using a dataset that doesn't fit in memory, using
tf.Dataset.shuffle will not work, because it requires reading the whole
thing into memory. With this solution, calling as_dataset with
shuffle=True will create the tf dataset with a generator that shuffles
the underlying index for each epoch. If the fits_in_memory option is
False, the shuffling will simply do nothing, as you can just use the
tf.shuffle instead.

This also changes as_numpy and as_dataset from properties to methods, so
watch out for that.