Refactors simple_nn. Might break experiments
I've refactored some of the ecg_cnn models a bit, which now also gives
some added flexibility for using resnet models and other pretrained
models. This alters the expected input arguments to all these models,
which will break experiments. I updated all the experiments which
I think were affected, but I haven't tested to make sure they run as
expected, so it might be reasonable to rerun them all just in case.

This also adds a new experiment for a pretrained resnet model which will
be the first part of a two-part model. In the first part, only the final
(non-pre-trained) layers are updated, and in the second part, the whole
thing (including the resnet part) is fine tuned.