Updates 1ECG hyper-experiment
Due to a previous, small refactoring of the ecg_cnn function, the input
kwargs in the hyper-experiment were wrong. I moved the ecg_ffnn_kwargs
from cnn_kwargs to one level up, and added an ecg_comb_ffnn_kwargs. The
structure of the experiment should be exactly the same.

Note that I didn't update the other hyper-experiments, so if I were to
re-run them, the same thing would have to be updated.