Adds new hyper-experiment
Hyper-experiment for resnet with 1 ecg. Also updates the M_F*
experiments, so that PCA is performed on the forberg features only and
not the flat-features. The sklearn_processor now expects a dict of
processors mapping each group of features to a specific processor.

The hyper-experiment adds a new functionality to the hyper-parameters:
SortedChoices, which works the same as Choices, except the resulting
list is sorted.

Also fixes mistake in the CNN7/CNN8 experiments, in which the ECGs are
not down-sampled before being fed into the network. All
M_R{1,2}_CNN{7,8}* experiments must therefore be re-run.

Finally, a CNN10 experiment is added.