Adds forberg hyperexperiments
Adds random-search hyper-experiment for the forberg features.
Modifies the resnet hyper-experiment (again), changing the dense sizes
slightly and adjusting the learning-rate scheduler to be more flexible.
It will now pick two random learning rates and train for 5-50 epochs
using the larger learning-rate before changing to the smaller one.

This also changes (again) the structure of the forberg-features, because
apparently the x-data can't be nested too deep (at least I think that's
the problem, but I didn't investigate it very much). The LR forberg
experiments need to be updated accordingly.