Refactors all the ESCT experiments
I decided to remove most of the legacy experiments (they were last
backed up in june, so it should be ok). I changed the naming convention
a bit and renamed the CNNs. The experiments had to be re-run anyway, so
I felt like this would be a bit cleaner, less chance for mistakes. All
experiments need to be re-run after this change, obviously.

Also adds a convenience function to the Presenter module, and some small
refactorings in the massage module.