Fixes the sklearn_preprocessor column names order
Fixes so that the order of the column names in the processed data is
correct. The order of the columns are currently changed if there are
categorical features present. That should maybe be fixed later, but for
now, at least the order of the names of the columns should be updated
propdata is correct. The order of the columns are currently changed if
there are categorical features present. That should maybe be fixed
later, but for now, at least the order of the names of the columns
should be updated properly.