Adjusts the ResNet tuning experiments
Changes number of epochs from 100 to 50, and the total number of
iterations from 400 to 200. Also adjusts the max cut-off for the
learning-rate reduction from 50 epochs to 25 epochs, to be more in line
with the reduction in total epochs.

The purpose of the change is to reduce computation time of the
experiment. This was done in the middle of training - after running 45,
49, 43 and 45 iterations for R1, R2, R1_FF and R2_FF respectively, to be
more in line with the reduction in total epochs.

The purpose of the change is to reduce computation time of the
experiment. This was done in the middle of training - after running 45,
49, 43 and 45 iterations for R1, R2, R1_FF and R2_FF respectively.