Adds function to make esc-trop cohort for article
In a review comment on the 1-TnT article, we were asked to validate the
results on a different dataset. Since that study was done on Expect, we
decided to try it on ESC-Trop. The code added in this commit extracts
the relevant data from ESC-Trop in a format suitable for Anders to do
the rest. May need some refactoring to look nice, but the
make_dataframe_for_anders function should return everything that is
needed.