Changes git-hook to use the pre-commit program
After updating my flake8 pre-commit stopped working. It seems that
flake8 now recommends using a program called pre-commit to manage all
pre-commit hooks. I therefore added it to the conda environment, and
updated the Makefile. Now instead of copying the .hooks/pre-commit file,
when setting up the project you have to do pre-commit install, so that
is added to the "make hooks" makefile.