Changes definition of MACE outcome for ESC-TROP
Breaking change! It turns out that a few unwanted intervention codes
where mistakenly included in the definition of MACE. Specifically the
ones that start with DG, corresponding to intubation and respirator
treatments should not have been included as part of MACE. In a meeting
with supervisors, it was decided to re-run the models with the new
target, but not to re-do the tuning process which is much more time
consuming.