Adds simple util to save/load with metadata
Save function that saves (pickles) data together with a metadata
report indicating timestamp, git commit, conda environment and so on.
Load function that loads the data at the given position, checks that the
metadata associated with the file is appropriate (for now, just that it
was generated without uncommitted changes), and returns the data.