Adds functions for multihot encoded diagnoses
Function to generate multihot encoded diagnoses for the Skåne 17-18
dataset, based on slutenvård and öppenvård. Each row in the resulting
dataframe corresponds to one visit, and indicates which diagnoses was
set at that time - a bool-vector of length 3174.

It takes a few minutes to make the multihot matrix, so I added
a function to save the whole thing as a sparsely encoded matrix and
pickle the result using the metadata.save util function.