Adds function to stagger diagnosis history
Given a matrix with one row for each diagnosis event per patient, create
a new matrix where each event is repeated for each chest-pain index
visit, as long as the index visit was after the diagnosis event. Thus,
a patient with multiple chest-pain visits will have their "history"
repeated, or "staggered". I add a function here to save this dataframe
to disk, as it takes some time to calculate. All this is in preparation
for the upcoming RNN experiments - it's pretty easy to turn this
staggered matrix into a RaggedTensor.