Adds new experiments trying random-split
The question of the train/val/test split came up once again, and it was
decided that I should re-run the final models using random split instead
of chronological split. This is the crude version -- I do a 50/50
stratified split, and repeat the models 10 times as before. A lot of
copy-pasta of Experiments - parameterizing stackers like this is now on
my wish-list.

It would have been easier to just do a 10-fold cross-validation instead,
but although that might have been better, it would have been a bit more
different from the previous tests...