Adds option to save model outside of callbacks
The validate model will now try to save the model if that option is
turned on (which it is by default). This works with the KerasWrapper
now, but not with sklearn models -- that's still TODO.

Also adds a convenience function to the Presenter class, to see the
final "history" of experiments.