Adds rhythm information to the ECG-hdf5 file
The ECG file contains, for each ECG, a list of "rhythm-strings",
describing various automatically detected rhythm-related pathologies,
such as atrial fibrillation or A-V block. I collected all possible
strings from all the ECG files and noted that they mostly were
combinations of statements separated by the word 'with', such as:
'Sinus bradycardia with sinus arrhythmia'. In order to compress this
information a bit and to also possibly make it easier to search in the
future, I now split all the rhythm statements into these sub-parts and
multi-hot encode the them into a vector of length 75. These are then
saved, similarly to the automated diagnoses, into the hdf5-file.

Note: the hdf5-file needs to be regenerated for these changes to be
accessible.