Adds DataWrapper and ExperimentResult
This is a major refactoring, with some old stuff not fully fixed. It
will break old experiments, or rather, old stuff must be rerun to be
compatible.

DataWrapper class that improves on the API for the Extractor class - it
is now more explicit what the expected output of the Extractor should
be, that it should include an x and a y, an index and names for the
columns. It still needs some work, but it's a step in the right
direction I think.

Experiment now produces a ExperimentResult dataclass as output, which
makes the code easier to read and should also make it easier to analyze
the results later on. Before, everything was just a plain old dict,
which has some advantages, but is also more difficult to read and
understand. Disadvantages with the new system is of course that new
experiment results will not be compatible with old ones, and the
Presenter class must be updated (not yet done).

I fixed (I think) most of the old experiments so that they should be
able to run with the new stuff. I didn't fix Anders' experiments, nor
the old troponin experiments (which I removed because I couldn't find
the data files anyway). I also didn't fix the ECG-autoencoder, but we
never ran that anyway, and it requires some more work I think before
streaming from disk will work properly with the new DataWrapper (since
it extends the old Data class that should now no longer be imported
directly I think).