Adds another exclusion criteria to esc-trop
Two patients with non-swedish background were found in ESC-TROP that
shouldn't be included. So I added an option (which is on by default)
that removes them.

This also changes the behavior of the ChronologicalSplit
cross-validation function, in a sneaky way (I should maybe parameterize
it later): if the requested split has a remainder, that data point will
go to the training set instead of the test set. In other words, when
calculating the size of the train/test split, train is an integer that
is rounded _up_, while test gets the rest. Before this change, train was
the result of a cast to int, which rounds down. The reason for this
change is because of the new patients that should be excluded: one of
them were previously included in my test-set, and I wanted the train/val
sets to remain the same as before. This was the simplest solution.