Changes to the run API
Adds a new input argument, action, which takes the value train or test.
Depending on the action, running an experiment will either train and
evaluate on the validation set as usual, or load a previously trained
model and evaluate on the test-set. The results of "testing" an
experiment will be saved to disk. This change should make it easier to
evaluate experiments on the test-set, if necessary.

The Extractor interface have replaced the get_data method with
a get_development_data (returning the training+validation set) and
a get_test_data (returning the held-out test set). It is up to the
implementer of the Extractor to make sure that these behave as expected.
The DDSA-extractor has been updated, but the others are TODO.

Refactoring of the ExperimentResult dataclasses. By default I will now
log the predictions on the training and validation sets after each
epoch, and save these in the training results. The other training
history is currently omitted. The intent is for the ExperimentResult
class to have convenience functions for easily calculating various
scores and metrics on the training/validation history.

After the training process, the finished model is saved to disk. Keras
models are saved as model.keras, and scikit-learn models are saved
(pickled) as model.sklearn. The models are automatically loaded when
evaulating on the test set.

Also adds a README to the serial_ecg project (WIP). Still need to update
the main README.