Adds cache module
Decorate a function or method with the cache.decorator.cache function,
and (depending on the cache settings) each time the function is called
with a given argument, its output will be written to disk. Subsequent
calls with the same argument will then load the results from disk
instead of actually calling the function, which can save time if the
function is expensive.

Caching in this way can be risky: if you change the function that was
cached (or any of its dependencies), or if you change your environment,
or if the function depends on the state of some other object outside of
that function (e.g. the function uses rng), any previously cached
outputs from that function will be 'wrong'. The cached results should
therefore be invalidated whenever there's a risk that the function has
changed. This is where the cache settings come in.

Along with the output of the function, some metadata is collected about
the state of the project at the time of caching: the commit hash,
if there are uncommitted changes to tracked files, the conda
environment, etc. A new metadata report is generated whenever the cached
function is called again, and the metadata is compared. Depending on the
cache settings, the cache may then be invalidated.

There are four settings:
Off - no caching is performed
Safe - cache is invalidated if either metadata report has uncommitted
changes in the code, if their commit has is different, or if the conda
environment is different.
Unsafe - Same as safe, but commits and conda environments may differ.
Stupid - Cache is never automatically invalidated.

Cache setting can be changed by calling the settings.set_level function
with the appropriate argument.

Only use the unsafe or stupid cache settings in low-stakes situations,
and at your own risk!