Bunch of refactoring and additional work on sk1718
Starts to move some common massag-efunctions to new module
Removes some old sk1718 massage code that was specific to
a pre-processed file from AB
Bunch of small changes to multihot-encoding process of sk1718 massage
New module for truncating or "rounding" ICD and ATC codes
New code for creating an index for all of sk1718
New code for creating MACE-like table of diagnoses for sk1718. Contains
some duplicated code from esc-trop that can probably be consolidated
into something more generalized later.

Fixes mistake in hold-out split for sk1718 extractor. This introduces
a new utility function for the DataWrapper class, which is a generalized
form of lazy-splits: input a list specifying the number of the split for
each element in the underlying data, and get the corresponding
partitioned dataset back.

Bunch of new and refactored experiments for the patient history project.