Refactorings
* Maps the values of the LISA variables to something more interpretable
* Groups the LISA variables into a few different categories, that can be
  selected individually from the Extractor
* Evaluating on test-set will now pre-process the data as needed. This
  uses the entire development set for the pre-processing step. Later,
I will want to refactor this so that pre-processing becomes part of the
model, and the pre-processing parameters learned on the training set
gets saved along with the model.
* Refactors get_cross_validation method in Experiment
* Adds 'ensemble' parameter to Experiment, which indicates how many
  times to repeat the entire cross-validation loop. This allows the
results to be interpreted as ensembles of identical models with
different random initialization.
* Refactors mlp_helper model. Simplifies the input arguments: instead of
  a dropouts and default_dropout, there's a single dropout argument that
is either a float (indicating the same dropout for each layer) or a list
of floats (one per layer). Similarly for the batch_norm and regularizer
parameter. The regularizer parameter is either a float, list of floats
or dict. In case of a dict, the keys 'activity', 'bias' and 'kernel'
must all be specified and map to a float or list of floats.
* Adds new mlp3 model for the PatientHistory project, which simply
  concatenates all inputs before feeding into a sequence of dense
layers.
* Adds a bunch of new experiments for the PatientHistory project.
* Removes and refactors a few PatientHistory experiments.
* Adjusts PatientHistory LR-experiments to tune the C-parameter.