WIP EnsembleRandomSearch
I wanted a new search strategy for hyper parameter optimization that
repeats each experiment up to k times, averaging the predictions from
each model. But this is expensive, and unnecessary for most experiments.
The idea with EnsembleRandomSearch strategy (WIP) is to only run one
iteration of each experiment, then filter out to top, say, 20% of those
and only run additional iterations for those. This should drastically
reduce the amount of compute resources, without affecting thte end
results. While doing this, I also intended to add some measure of how
jagged the loss-history is, and penalize models that have poor
performance in this regard, so as to only fully explore models that are
well behaved. But this jaggedness measure is not implemented yet.

In preparation for the new search strategy, I added changes that allow
for an experiment to save partial results after each cross-validation
fold. Using the --rerun option will restart partial results from
scratch, otherwise any half-finished experiment should pick up where it
left off.

This commit also modifies slightly how the Result class deals with
ensembled predictions, but this shouldn't affect how the results
themselves are actually saved, and so should be backwards compatible.