WIP stupid implementation of Mattias' idea
This breaks a lot of stuff. Idea is to override the tensorflow Model
class method train_step, so that the sample_weights becomes a model
attribute. This attribute can then be dynamically updated by a callback
after each epoch. The callback and new model are now hard-coded into the
project just to see if it works.

In order for the sample_weights to work properly, the class attribute
must be sliced, and right now I just add the index where the
sample_weights would otherwise be, and try to reindex the sample_weights
according to this (because each batch gets randomly shuffled). But it is
unreasonably difficult to properly slice a tf.Tensor, and that's where
the code breaks currently.