Adds ability to resume experiments
If the -r (rerun) flag is off, we now check if there are any saved
models in the checkpoints path, and if so, the experiment is treated as
partially completed. The model is then loaded from the latest checkpoint
and training is resumed where the experiment was interrupted.

TODO: The ExperimentResult object is currently not saved, either between
splits or during loading. There is also a bunch of refactorings to do
there. There is also no check right now for metadata compatibility,
since I still haven't figured out how to deal with micromamba
environments properly.