Fixes the metadata environment logging
Now the metadata report can log installed packages of the current
environment, if the user has micromamba installed.

Changes how the cache-files are written: now the metadata file is
written first, and when loading a cache, we first check that the
metadata is valid before attempting to load the actual payload. This
way, we can reject a large invalidated cache without actually trying to
load it into memory.

Also changes the caching to be turned off by default. Use -c
safe/unsafe/stupid to turn it on.