Fixes PredictionLogger & new experiments
The PredictionLogger was broken with the new version of tensorflow,
since for some reason it can now only store scalars in the logs dict.
But now I've finally implemented a (ugly) workaround for saving the
predictions during training.

Also adds new target experiments checking the performance of RN1 + SE8,
as well as the pre-training predicting age and sex simultaneously.

New source tasks include a small grid-search for the reduction ratio of
the squeeze and excitation block.