Fixes failing tests
The test_run experiment needed a small change due to the API change of
the optimizer parameter in the Experiment class.
The test_experiments failed because the modelwrapper required to know
the length of the training set, also because of the API change to the
optimizer/learning_rate parameters.

Also a small, likely temporary experiment to check the overfitting
behavior of the PTAS experiment.