Changes to Experiment API & new experiments
Caution, this change breaks old experiments! Now, instead of providing
extractor_kwargs as a single dict in the Experiment, specify separately
the extractor_index, extractor_features and extractor_labels. Although
it breaks old experiments (ie you have to go back and edit the
extractor_kwargs to fix it), it is more convenient to repeat experiments
that only adjusts some of the extractor_kwargs. Use the
data_fits_in_memory Experiment attribute to specify if data fits in
memory. This parameter is automatically sent to the Extractor, though it
is still up to the implementation of the Extractor class to make use of
that parameter.

This commit also fixes all experiments in the transfer project to comply
with the new API change. Also retrospectively updates calls to
TargetExtractor, as well as adding save_val_prediction_history flag to
all experiments.

Also updates the names of the transfer experiments to more consistently
indicate pre-training labels: PT_RN1_R100 becomes PTS100_RN1_R100,
meaning it was pre-trained on sex, using 100% of the data.

Finally, adds some new Target experiments, although the ones using RN2
are still waiting for the pre-training to complete.