Updates to the OMI project
The omi-table now considers all the scaar entries for each patient,
instead of just the first one. For all the relevant columns, it is
enough to have one hit in any of the scaar entries, for that column to
be marked as True.

Also updates the annotation documents in the omi extractor, as well as
some small refactorings.