Breaking changes in transfer project
Fixes mistake in temporal_split of the transfer extractor, causing data
points to be erroneously excluded. Shouldn't affect results so far, but
the fix will require all experiments in the transfer project to be
re-run. This fix also changes slightly how the split is calculated in
the first place: now will take as many of the last visits
(chronologically) as needed to include exactly 15% of the patients, and
use that as the temporal test split (as opposed to taking 15% of the
visits). The rest should proceed as before.

Makes the temporal/random test splits explicit in the test-set returned
by the extractor.

DataWrapper to_dataframe now correctly uses MultiIndex when the index
has multiple columns (as in the test-split of the transfer project).

Adds experiments testing the predicted age/sex as input features to
logistic regression.

Fixes wrong batch-size on one of the grid-search experiments.

Adds transfer project latex template and WIP manuscript.