Changes optimizer of fine-tuning experiments
With the new version of tensorflow it seems that I need to use the
legacy version of the Adam optimizer if I want to unfreeze models the
way I have before. So I fixed that.

Also adds some missing fine-tuning experiments and fixes a typo in the
CNN fine-tuning experiment.